--- mealkitary-application/src/main/kotlin/com/mealkitary/demo/application/port/in/GreetUserQuery.kt
@@ -1,5 +1,0 @@
-package com.mealkitary.demo.application.port.`in`
-
-interface GreetUserQuery {
-    fun execute(userId: Long): GreetUserResponse
-}
