--- mealkitary-application/src/main/kotlin/com/mealkitary/demo/application/port/in/GreetUserResponse.kt
@@ -1,5 +1,0 @@
-package com.mealkitary.demo.application.port.`in`
-
-data class GreetUserResponse(
-    val message: String,
-)
