--- mealkitary-domain/src/main/kotlin/com/mealkitary/domain/reservation/ReservationLineItem.kt
@@ -1,3 +1,0 @@
-package com.mealkitary.domain.reservation
-
-class ReservationLineItem
