--- mealkitary-domain/src/main/kotlin/com/mealkitary/domain/reservation/ReservationStatus.kt
@@ -1,24 +1,0 @@
-package com.mealkitary.domain.reservation
-
-enum class ReservationStatus {
-    NOTPAID,
-    PAID,
-    RESERVED,
-    REJECTED, ;
-
-    fun isNotPaid(): Boolean {
-        return this.equals(NOTPAID)
-    }
-
-    fun isReserved(): Boolean {
-        return this.equals(RESERVED)
-    }
-
-    fun isRejected(): Boolean {
-        return this.equals(REJECTED)
-    }
-
-    fun isPaid(): Boolean {
-        return this.equals(PAID)
-    }
-}
