--- mealkitary-domain/src/main/kotlin/com/mealkitary/domain/shop/ShopStatus.kt
@@ -1,10 +1,0 @@
-package com.mealkitary.domain.shop
-
-enum class ShopStatus {
-    INVALID,
-    VALID;
-
-    fun isInvalidStatus(): Boolean {
-        return this.equals(INVALID)
-    }
-}
