+++ mealkitary-domain/src/main/kotlin/com/mealkitary/reservation/domain/ReservationLineItem.kt
@@ -1,0 +1,3 @@
+package com.mealkitary.reservation.domain
+
+class ReservationLineItem
