+++ mealkitary-domain/src/main/kotlin/com/mealkitary/shop/domain/ShopStatus.kt
@@ -1,0 +1,10 @@
+package com.mealkitary.shop.domain
+
+enum class ShopStatus {
+    INVALID,
+    VALID;
+
+    fun isInvalidStatus(): Boolean {
+        return this.equals(INVALID)
+    }
+}
